+++ lib/service/Config.dart
@@ -176,6 +176,12 @@ class Config extends rokwire.Config {
   int  get recentItemsCount          => JsonUtils.intValue(settings['recentItemsCount']) ?? 32;
   int  get homeRecentNotificationsCount => JsonUtils.intValue(settings['homeRecentNotificationsCount']) ?? 5;
   int  get homeRecentPollsCount      => JsonUtils.intValue(settings['homeRecentPollsCount']) ?? 3;
+  
+  int  get appReviewSessionsCount    => JsonUtils.intValue(settings['appReviewSessionsCount']) ?? 24;
+  int  get appReviewSessionDuration  => JsonUtils.intValue(settings['appReviewSessionDuration']) ?? 120;
+  int  get appReviewActivityTimeout  => JsonUtils.intValue(settings['appReviewActivityTimeout']) ?? 3;
+  int  get appReviewRequestTimeout   => JsonUtils.intValue(settings['appReviewRequestTimeout']) ?? 21;
+  
   String get appPrivacyVersion       => JsonUtils.stringValue(settings['privacyVersion']) ?? (JsonUtils.stringValue(content['mobileAppVersion']) ?? '0.0.0');
   bool get allowGroupsAuthmanSync    => JsonUtils.boolValue(settings['allowGroupsAuthmanSync']) ?? false;